Added a CombinedStreamGroup to simplify extension for new devices
